fel: sid: add SID dump maps for other SoCs
At the moment we only have a SID map for the R40. Add a generic
placeholder map, covering the 2048 bits that most SoCs have (at least).

The linux-sunxi Wiki page [1] for the SID register layout describes
entries for the H6 and "SoCs before H6". Add those tables as well, and
use the latter one for some SoCs which were released the closest to the
H6 (H3, A64, H5).
If people have more information, more specific maps can be added.

[1] https://linux-sunxi.org/SID_Register_Guide#eFUSE_Region_Overview

Signed-off-by: Andre Przywara <osp@andrep.de>