uart0-helloworld-sdboot: make remaining function static
Our uart0-helloworld-sdboot tools lives in a single source file, so we
don't need to export any functions we use. Nevertheless most functions
were in the past not marked as "static".

Mark the one function that hasn't been touched and updated in the last few
patches as "static", to avoid the compiler exporting them, and to help
it with inlining functions.

Signed-off-by: Andre Przywara <osp@andrep.de>