uart0-helloworld-sdboot: add clock data to the SoC table
To make the UART work, we have to open the clock gate and de-assert the
reset for the UART0 peripheral. This is currently done by calling one of
four functions, even though there are actually only two different ways:
- Older SoCs need to set a bit in the APB1/2 bus clock register, and the
  same bit in the respective reset register.
- Newer SoCs combines both the clock and the reset gate in one register,
  so it's only one write, setting two bits.
Add a flag to the table to choose between the two methods for each SoC.
The offsets for these clock registers are very stable, but the CCU base
address changed a bit over time, so split this up and put the base address
separately into the table.

For the latest SoCs, the "new" clock structure is combined with an
updated GPIO IP, and the Allwinner BSP calls this SoC generation
"NCAT2". Add a symbolic name for this already, so we can mark those SoCs
now, and add the respective GPIO flag later when this gets added.

Signed-off-by: Andre Przywara <osp@andrep.de>