uart0-helloworld-sdboot: add GPIO data to the SoC table
To make the UART signal appear on the SoC pins, we have to select two
pins from the supported output pins and set up their pinmux to connect
to the UART peripheral.
As the name suggests, uart0-helloworld-sdboot always uses UART0, which
is used as a debug UART on most boards, and is also available as a
multiplex on the SD card pins, to be used with a breakout board.

At the moment there is a large if/else cascade, with each branch calling
three GPIO setup functions, with the right arguments for each SoC.
While this is easily extendable with copy&paste, it is very repetetive
and duplicates quite some code.

Use our new SoC table to put the pin number for the TX pin and the
required pinmux value in there, alongside the PIO base address. The RX
pin is so far always the pin after the TX pin, so we just need to store
one pin. We also store a flag marking the PIO IP generation for each
SoC, so we can just differentiate this easily in the GPIO setup
functions.
This also uses the opportunity to get rid of the idea to assign some
names to the pinmux values, when they are just seemingly random SoC
specific numbers. Instead just use a macro name with that number in it,
to make it obvious that this is the mux value, but don't try to put some
deeper meaning into them.

On one SoC we decided to use the SD card pinmux, which is always on pins
PF2 and PF4. Add a flag to allow a SoC to default to those pins.

Also remove the unneeded GPIO functions, and mark the rest of them
as static.

Signed-off-by: Andre Przywara <osp@andrep.de>