uart0-helloworld-sdboot: add SRAM A1 base address to SoC table
After we have printed our greeting message, the code tries to figure out
from which media the devices has been booted, and prints that
information. When we were booted via FEL, we return there, so further
code can be uploaded.

To access this boot source information, we need to read from offset 0x28
of SRAM A1. The SRAM base address is different between the SoC
generations, so add this address to the table and use it directly.

Signed-off-by: Andre Przywara <osp@andrep.de>