uart0-helloworld-sdboot: remove old soc_is_xxx() macros
Now that every SoC specific aspect of our tool is covered by the SoC
table, we don't need the old macros testing for each SoC anymore.

Remove them alongside the no longer needed versions to differentiate
the few SoC variants, which are already covered by a flag in the table.

Signed-off-by: Andre Przywara <osp@andrep.de>