uart0-helloworld-sdboot: add support for Allwinner A523
The Allwinner A523 is an SoC with eight Arm Cortex-A55 cores, and some
updated peripherals. The basic IP is very close to other recent SoCs, so
we just have to add the respective flags and base addresses to the new
SoC table, to allow boards with this SoC to run uart0-helloworld-sdboot.

The A523 comes with a GICv3, at a different address, so add this to the
early generation detection code, and let it use the H6 SRAM controller
address for further SoC-ID detection.

Tested on an Avaota A1 board.

Signed-off-by: Andre Przywara <osp@andrep.de>