uart0-helloworld-sdboot: add support for Allwinner A80
The Allwinner A80 is quite an oddball: it uses a big.LITTLE configuration
with Arm Cortex-A7 and A15 cores, and has a completely different memory
map compared to all the other Allwinner SoCs.

Add support for this SoC, by adding the respective base addresses for its
peripherals and the pinmux into our SoC table.
It also requires a small extension to the SoC detection routine, to
accommodate for the different GIC mapping.

Tested on an Cubieboard 4 (CC-A80).

Signed-off-by: Andre Przywara <osp@andrep.de>