spi: use GPIO and CCU base address from soc_info
Now that we store the base addresses of the GPIO and clock controllers
for each SoC in our soc table, let's use those values instead of
determining them ourselves, in the SPI flash code.
This involves several changes:
- removing the base offset from each register address
- supporting the NCAT2 pin controller, with a larger per-port size
- using info about clock and pinctrl generation from soc table

This will simplify future SoC support, as we can simply put the
respective base addresses in the soc table, without touching the
SPI flash code.

Signed-off-by: Andre Przywara <osp@andrep.de>