spi: move SPI0 base address into soc_info
Use the main soc table to also store the first SPI controller's MMIO
base address, so that we don't need to decide this in the SPI flash
code.
Only SoCs that are covered by fel-spiflash.c get a base address assigned,
so this value being not 0 is used as an indicator that the SoC is
supported.

This will simplify supporting future generations of SoCs.

Signed-off-by: Andre Przywara <osp@andrep.de>