spi: use pins and pinmux data from soc_info
The bootable SPI flash is always connected to PortC pins, though the
exact pin numbers within PortC and their pinmux differs between the SoCs.

We always need four pins (MOSI, MISO, CLK, CS), so pack their pin numbers
into one 32-bit word and store that in our soc table. Also store the
required pinmux (which is always the same for all four pins) in the table.
Then use this information in the SPI flash code to extract the pins again
and configure them accordingly.
This does away with the fragile switch/case construct we used to configure
the pins so far.

Signed-off-by: Andre Przywara <osp@andrep.de>